Move logging tests somewhere more reasonable.
